test(swa): consolidate SWA tests, drop bespoke runner + mirror-model test
Cohesion pass over the SWA test surface (no coverage lost except the one test
that contradicts the node-mount architecture):

- DELETE test_kvmanager_dsv4_layerwise_swa.py — it validates the REJECTED SWA
  mirror model (get_match + launch, swa_verifier expects SWA@full-KV-block).
  Incompatible with node-mount by design (SWA rides get_match_swa, not full-KV).

- MERGE node-mount invariant tests into one file, test_swa_node_mount.py, with
  three sections: (1) pure-Python RadixTreeIndex spec, (2) C++ CRadixTreeIndex
  production (was test_swa_accel_nodemount.py), (3) a stochastic multi-turn
  workload stress test folded in from benchmarks/benchmark_swa_nodemount.py
  (pressure-driven I1/I2 + real SWA byte-verify + no-leak-after-reset), now a
  parametrized pytest with small CI-sized params. Delete the standalone benchmark.

- MERGE test_swa_control_plane_smoke.py + test_swa_launch_bind.py ->
  test_swa_control_plane.py (same subsystem: graph build + late-bind, shared
  helpers deduped).

- DROP tests/run_test_plan.sh. pytest already does discovery + reporting; the sh
  was a hand-maintained file list that went stale and mixed pytest with script
  invocations. Replaced by pytest markers (unit/smoke/e2e) registered in
  pyproject.toml, plus testpaths=["tests"] so collection never wanders into
  third_party/ googletest .py files (the reason the sh used an explicit list).
  The two GPU e2e scripts (test_swa_dispatch.py, test_swa_control_plane_e2e.py)
  gain a thin @pytest.mark.e2e wrapper so `pytest -m e2e` collects them while
  they stay runnable standalone.

Verified: pytest -m unit = 64 passed/1 skipped; -m smoke = 8 passed/1 skipped;
bare collection = 430 tests, no third_party errors.
