+++ tests/test_swa_node_mount.py
@@ -1,23 +1,50 @@
 # SPDX-FileCopyrightText: Copyright (c) <2025> NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
-"""Node-mounted SWA logic tests for the pure-Python RadixTreeIndex mirror.
+"""Node-mounted SWA on the radix tree — invariants I0-I4 + a workload stress run.
 
-These validate the node-mount re-architecture (deployments/swa_design/
-08_节点挂载SWA架构.md): SWA mounted on radix nodes, split-preserve, unified
-match, dual eviction (full-by-leaf connects SWA; SWA-only LRU), dual lock.
+The node-mount re-architecture (deployments/swa_design/02_控制面_节点挂载匹配与驱逐.md)
+mounts SWA on radix nodes. This file is the single home for its invariant
+coverage, in three sections:
 
-The pure-Python RadixTreeIndex (flexkv/cache/radixtree.py) is the executable
-spec for the C++ CRadixTreeIndex; DSv4 runs the C++ path (FLEXKV_INDEX_ACCEL=1).
-Runs without torch/GPU (numpy only), so the invariants are checkable locally.
+  1. PURE-PYTHON SPEC  (flexkv/cache/radixtree.py RadixTreeIndex) — the
+     executable spec; numpy-only, runs anywhere.
+  2. C++ PRODUCTION    (flexkv.c_ext CRadixTreeIndex, the DSv4 path with
+     FLEXKV_INDEX_ACCEL=1) — the same invariants on the shipped implementation.
+     (test_swa_cnode_cascade.py targets the P2P-only LocalRadixTree and skips in
+     the default build; this section is the production coverage.)
+  3. WORKLOAD STRESS   — a stochastic multi-turn dialogue through the production
+     CacheEngineAccel path under pool pressure, with REAL SWA host-pool byte IO,
+     asserting no double-free / use-after-free / byte-mismatch / leak-after-reset
+     (folded in from the former benchmarks/benchmark_swa_nodemount.py).
+
+Invariants:
+  I0  each node holds ≤1 SWA at its trailing page; split preserves it on the
+      suffix half; merge moves it to follow the merged node's last page.
+  I1  SWA ⊆ Full: freeing a node's Full KV frees its SWA slot (drained to pool).
+  I2  SWA-only eviction prefers interior nodes; a leaf that loses its SWA with no
+      full lock is deleted, with a full lock its Full KV is kept.
+  I3  full_lock_ref (lock_cnt) ≥ swa_lock_ref, symmetric inc/dec.
+  I4  match returns the deepest fully-matched ready node with a live SWA in one
+      pass; a partial node match must not expose its trailing-page SWA.
 """
+import random
+from dataclasses import dataclass, field
+from typing import Dict, List
+
 import numpy as np
 import pytest
 
 from flexkv.common.block import SequenceMeta
 from flexkv.cache.radixtree import RadixTreeIndex
 
+pytestmark = pytest.mark.unit
+
 TPB = 2
 
+
+# =========================================================================== #
+# 1. PURE-PYTHON SPEC — RadixTreeIndex (numpy only)                           #
+# =========================================================================== #
 
 def _seq(ids):
     return SequenceMeta(np.array(ids, dtype=np.int64), tokens_per_block=TPB)
@@ -27,7 +54,7 @@ def _phys(*xs):
     return np.array(xs, dtype=np.int64)
 
 
-def test_match_returns_last_swa_node():
+def test_py_match_returns_last_swa_node():
     idx = RadixTreeIndex(tokens_per_block=TPB)
     n1 = idx.insert(_seq([1, 2, 3, 4, 5, 6, 7, 8]), _phys(0, 1, 2, 3), is_ready=True)
     assert n1 is not None and n1.size() == 4
@@ -40,7 +67,7 @@ def test_match_returns_last_swa_node():
     assert mr.swa_hit_blocks == 4
 
 
-def test_split_preserves_swa_on_suffix_half():
+def test_py_split_preserves_swa_on_suffix_half():
     """I0/I4: split keeps the SWA on the half that owns the original last page."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
     n1 = idx.insert(_seq([1, 2, 3, 4, 5, 6, 7, 8]), _phys(0, 1, 2, 3), is_ready=True)
@@ -61,7 +88,7 @@ def test_split_preserves_swa_on_suffix_half():
     assert mr.last_swa_node is n1 and mr.swa_hit_blocks == 4
 
 
-def test_full_evict_frees_swa_slot():
+def test_py_full_evict_frees_swa_slot():
     """I1: evicting a node's Full KV frees its SWA slot (drained to pool)."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
     na = idx.insert(_seq([1, 2, 3, 4]), _phys(0, 1), is_ready=True)
@@ -73,7 +100,7 @@ def test_full_evict_frees_swa_slot():
     assert idx.total_swa_slots() == 0
 
 
-def test_evict_swa_prefers_internal_node():
+def test_py_evict_swa_prefers_internal_node():
     """Multi-turn: SWA-only eviction drops interior-prefix SWA first, keeps Full."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
     nfull = idx.insert(_seq([1, 2, 3, 4, 5, 6, 7, 8]), _phys(0, 1, 2, 3), is_ready=True)
@@ -95,7 +122,7 @@ def test_evict_swa_prefers_internal_node():
     assert evf.size == 0  # internal SWA evict frees no full blocks
 
 
-def test_evict_swa_leaf_without_lock_deletes_node():
+def test_py_evict_swa_leaf_without_lock_deletes_node():
     """I2: a leaf that would lose its SWA and has no full lock is deleted whole."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
     nl = idx.insert(_seq([1, 2, 3, 4]), _phys(0, 1), is_ready=True)
@@ -107,7 +134,7 @@ def test_evict_swa_leaf_without_lock_deletes_node():
     assert idx.drain_freed_swa_slots() == [500]
 
 
-def test_evict_swa_leaf_with_full_lock_keeps_full():
+def test_py_evict_swa_leaf_with_full_lock_keeps_full():
     idx = RadixTreeIndex(tokens_per_block=TPB)
     nl = idx.insert(_seq([1, 2, 3, 4]), _phys(0, 1), is_ready=True)
     idx.set_swa(nl, slot=600)
@@ -118,7 +145,7 @@ def test_evict_swa_leaf_with_full_lock_keeps_full():
     assert not idx.is_empty()
 
 
-def test_dual_lock_invariant():
+def test_py_dual_lock_invariant():
     """I3: full_lock_ref (lock_cnt) >= swa_lock_ref, with paired inc/dec."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
     n1 = idx.insert(_seq([1, 2, 3, 4]), _phys(0, 1), is_ready=True)
@@ -130,7 +157,7 @@ def test_dual_lock_invariant():
     assert n1.lock_cnt == 0 and n1.swa_lock_ref == 0
 
 
-def test_dec_swa_lock_only_early_release():
+def test_py_dec_swa_lock_only_early_release():
     idx = RadixTreeIndex(tokens_per_block=TPB)
     n1 = idx.insert(_seq([1, 2, 3, 4]), _phys(0, 1), is_ready=True)
     idx.set_swa(n1, slot=701)
@@ -143,7 +170,7 @@ def test_dec_swa_lock_only_early_release():
     assert n1.lock_cnt == 0
 
 
-def test_dual_lock_only_deepest_swa_node_locked():
+def test_py_dual_lock_only_deepest_swa_node_locked():
     """I3 + scope: inc_lock_ref locks full on [node,root) but SWA only on the
     single deepest node with SWA; dec is symmetric (no underflow)."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
@@ -162,7 +189,7 @@ def test_dual_lock_only_deepest_swa_node_locked():
     assert a.lock_cnt == 0 and A.lock_cnt == 0
 
 
-def test_swa_locked_node_not_full_evictable():
+def test_py_swa_locked_node_not_full_evictable():
     """in_use() includes swa_lock_ref: a SWA-locked node is not full-evictable."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
     n = idx.insert(_seq([1, 2, 3, 4]), _phys(0, 1), is_ready=True)
@@ -172,7 +199,7 @@ def test_swa_locked_node_not_full_evictable():
     assert not n.evictable()
 
 
-def test_reset_rearms_swa_pool_via_host_pool():
+def test_py_reset_rearms_swa_pool_via_host_pool():
     """SWAHostPool.reset re-arms every slot free (tree reset drops all nodes)."""
     from flexkv.swa.swa_host_pool import SWAHostPool
     from flexkv.common.config import SWAPoolConfig
@@ -185,7 +212,7 @@ def test_reset_rearms_swa_pool_via_host_pool():
     assert pool.num_free == 4  # all slots reclaimed
 
 
-def test_partial_node_match_does_not_report_swa():
+def test_py_partial_node_match_does_not_report_swa():
     """§5.2: a partially-matched node does not expose its trailing-page SWA."""
     idx = RadixTreeIndex(tokens_per_block=TPB)
     n1 = idx.insert(_seq([1, 2, 3, 4, 5, 6, 7, 8]), _phys(0, 1, 2, 3), is_ready=True)
@@ -197,19 +224,9 @@ def test_partial_node_match_does_not_report_swa():
     assert mr.last_swa_node is None and mr.swa_hit_blocks == 0
 
 
-def test_merge_child_moves_swa_to_merged_node():
+def test_py_merge_child_moves_swa_to_merged_node():
     """I0: merging a single child moves the SWA to follow the child's last page,
     freeing the parent's stale SWA."""
-    idx = RadixTreeIndex(tokens_per_block=TPB)
-    # Build parent(2 blk, SWA=900) -> child(2 blk, SWA=901) by split.
-    n = idx.insert(_seq([1, 2, 3, 4, 5, 6, 7, 8]), _phys(0, 1, 2, 3), is_ready=True)
-    idx.set_swa(n, slot=901)  # this node is the 4-block leaf
-    sd = _seq([1, 2, 3, 4, 55, 66, 77, 88])
-    idx.insert(sd, _phys(8, 9), is_ready=True, match_result=idx.match_prefix(sd))
-    parent = n.parent  # prefix half (2 blk), give it a stale SWA
-    idx.set_swa(parent, slot=900)
-    assert parent.num_children() == 2  # n + the divergent node -> can't merge yet
-
     # Fresh 2-level chain we can merge: root -> A(2blk,SWA) -> B(2blk,SWA), A has 1 child.
     idx2 = RadixTreeIndex(tokens_per_block=TPB)
     a = idx2.insert(_seq([1, 2, 3, 4, 5, 6, 7, 8]), _phys(0, 1, 2, 3), is_ready=True)
@@ -227,3 +244,418 @@ def test_merge_child_moves_swa_to_merged_node():
     # and A's stale 900 was freed.
     assert A.swa_host_slot == 901 and not A.swa_tombstone
     assert 900 in idx2.drain_freed_swa_slots()
+
+
+# =========================================================================== #
+# 2. C++ PRODUCTION — CRadixTreeIndex                                         #
+# =========================================================================== #
+
+c_ext = pytest.importorskip("flexkv.c_ext")
+
+_CEXT_OK = "swa_host_slot" in dir(c_ext.CRadixNode) and all(
+    n in dir(c_ext.CRadixNode) for n in ("is_leaf", "get_lock_cnt", "has_swa")
+)
+_cext_reason = "CRadixNode SWA/structural bindings not present (rebuild c_ext)"
+cext = pytest.mark.skipif(not _CEXT_OK, reason=_cext_reason)
+
+import torch  # noqa: E402 — only needed for the c_ext section
+
+
+def _hashes(ids):
+    seq = SequenceMeta(token_ids=np.array(ids, dtype=np.int64), tokens_per_block=TPB)
+    return torch.from_numpy(seq.block_hashes.astype(np.int64))
+
+
+def _tree():
+    return c_ext.CRadixTreeIndex(TPB, 4096, 0, "lru")
+
+
+def _insert(tree, ids, base, ready=True, match=None):
+    n = len(ids) // TPB
+    bh = _hashes(ids)
+    phys = torch.arange(base, base + n, dtype=torch.int64)
+    if match is None:
+        return tree.insert(phys, bh, n, -1, ready)
+    return tree.insert(phys, bh, n, -1, ready,
+                       match.last_node, match.num_matched_blocks,
+                       match.last_node_matched_length)
+
+
+def _evict_full(tree, k):
+    buf = torch.zeros(k, dtype=torch.int64)
+    got = tree.evict(buf, k)
+    return buf.numpy()[:got]
+
+
+def _evict_swa(tree, k):
+    buf = torch.zeros(0, dtype=torch.int64)
+    freed = tree.evict_swa(buf, k)
+    return buf.numpy(), freed
+
+
+@cext
+def test_cpp_match_reports_last_swa_node():
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4, 5, 6, 7, 8], 0)
+    t.set_swa(n, 100)
+    assert n.has_swa() and n.swa_host_slot == 100
+    mr = t.match_prefix(_hashes([1, 2, 3, 4, 5, 6, 7, 8]), 4, False)
+    assert mr.num_matched_blocks == 4
+    assert mr.last_swa_node is not None
+    assert mr.last_swa_node.swa_host_slot == 100
+    assert mr.swa_hit_blocks == 4
+
+
+@cext
+def test_cpp_partial_node_match_hides_tail_swa():
+    """I4: SWA sits on the node's last page; a partial match cannot claim it."""
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4, 5, 6, 7, 8], 0)
+    t.set_swa(n, 800)
+    mr = t.match_prefix(_hashes([1, 2, 3, 4, 5, 6, 99, 98]), 4, False)
+    assert mr.num_matched_blocks == 3
+    assert mr.last_swa_node is None
+    assert mr.swa_hit_blocks == 0
+
+
+@cext
+def test_cpp_split_preserves_swa_on_suffix():
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4, 5, 6, 7, 8], 0)
+    t.set_swa(n, 200)
+    # Diverge after 2 blocks -> split the 4-block node into prefix(2)+suffix(2).
+    s2 = [1, 2, 3, 4, 55, 66, 77, 88]
+    m = t.match_prefix(_hashes(s2), 4, False)
+    assert m.num_matched_blocks == 2
+    _insert(t, s2, 8, match=m)
+    # original node is now the suffix half and KEEPS its slot; nothing freed.
+    assert n.swa_host_slot == 200 and not n.swa_tombstone and n.size() == 2
+    parent = n.parent
+    assert parent is not None and parent.swa_host_slot == -1 and parent.swa_tombstone
+    assert t.drain_freed_swa_slots() == []
+    mr = t.match_prefix(_hashes([1, 2, 3, 4, 5, 6, 7, 8]), 4, False)
+    assert mr.last_swa_node is not None and mr.swa_hit_blocks == 4
+
+
+@cext
+def test_cpp_full_evict_frees_swa_slot():
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4], 0)
+    t.set_swa(n, 300)
+    ev = _evict_full(t, 2)
+    assert len(ev) == 2
+    assert 300 in t.drain_freed_swa_slots()
+    assert t.is_empty()
+
+
+@cext
+def test_cpp_full_evict_no_swa_no_drain():
+    t = _tree()
+    _insert(t, [1, 2, 3, 4], 0)
+    _evict_full(t, 2)
+    assert t.drain_freed_swa_slots() == []
+
+
+@cext
+def test_cpp_evict_swa_internal_first_keeps_full():
+    """Multi-turn: interior-prefix SWA dropped first, its Full KV kept."""
+    t = _tree()
+    nfull = _insert(t, [1, 2, 3, 4, 5, 6, 7, 8], 0)
+    t.set_swa(nfull, 400)
+    sd = [1, 2, 3, 4, 99, 98, 97, 96]
+    m = t.match_prefix(_hashes(sd), 4, False)
+    _insert(t, sd, 8, match=m)
+    A = nfull.parent          # internal prefix node
+    assert A is not None and not A.is_leaf()
+    t.set_swa(A, 401)         # give the interior node an SWA
+    # Touch the leaf so the interior node A is the LRU victim.
+    t.set_swa(nfull, 400)     # re-mount == MRU bump on nfull
+    evf, nfreed = _evict_swa(t, 1)
+    assert nfreed == 1
+    assert 401 in t.drain_freed_swa_slots()
+    assert A.swa_tombstone and A.swa_host_slot == -1
+    assert A.size() == 2 and not A.is_leaf()   # Full KV kept
+    assert nfull.has_swa()
+    assert evf.size == 0                        # interior evict frees no full
+
+
+@cext
+def test_cpp_evict_swa_leaf_unlocked_deletes_node():
+    t = _tree()
+    nl = _insert(t, [1, 2, 3, 4], 0)
+    t.set_swa(nl, 500)
+    evf, nfreed = _evict_swa(t, 1)
+    assert nfreed == 1
+    assert evf.size == 2                 # whole node deleted, full freed
+    assert t.is_empty()
+    assert 500 in t.drain_freed_swa_slots()
+
+
+@cext
+def test_cpp_evict_swa_leaf_locked_keeps_full():
+    t = _tree()
+    nl = _insert(t, [1, 2, 3, 4], 0)
+    t.set_swa(nl, 600)
+    t.lock(nl)                           # full lock
+    evf, nfreed = _evict_swa(t, 1)
+    assert nfreed == 1 and evf.size == 0
+    assert nl.swa_tombstone and nl.size() == 2
+    assert not t.is_empty()
+
+
+@cext
+def test_cpp_dual_lock_symmetric():
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4], 0)
+    t.set_swa(n, 700)
+    b = t.inc_lock_ref(n)
+    assert b is not None
+    assert n.get_lock_cnt() == 1 and n.swa_lock_ref == 1
+    assert n.get_lock_cnt() >= n.swa_lock_ref     # I3
+    t.dec_lock_ref(n, b, False)
+    assert n.get_lock_cnt() == 0 and n.swa_lock_ref == 0
+
+
+@cext
+def test_cpp_only_deepest_swa_node_locked():
+    t = _tree()
+    a = _insert(t, [1, 2, 3, 4, 5, 6, 7, 8], 0)
+    t.set_swa(a, 10)
+    sd = [1, 2, 3, 4, 55, 66, 77, 88]
+    m = t.match_prefix(_hashes(sd), 4, False)
+    _insert(t, sd, 8, match=m)
+    A = a.parent
+    t.set_swa(A, 11)             # both interior A and leaf a carry SWA
+    b = t.inc_lock_ref(a)
+    assert b is not None
+    # deepest SWA node (the leaf a) is the boundary; only it is SWA-locked.
+    assert a.swa_lock_ref == 1 and A.swa_lock_ref == 0
+    assert a.get_lock_cnt() == 1 and A.get_lock_cnt() == 1   # full on both
+    t.dec_lock_ref(a, b, False)
+    assert a.swa_lock_ref == 0 and A.swa_lock_ref == 0
+    assert a.get_lock_cnt() == 0 and A.get_lock_cnt() == 0
+
+
+@cext
+def test_cpp_dec_swa_lock_only_early_release():
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4], 0)
+    t.set_swa(n, 701)
+    b = t.inc_lock_ref(n)
+    t.dec_swa_lock_only(b)
+    assert n.swa_lock_ref == 0 and n.swa_tombstone   # leaf SWA freed early
+    assert n.get_lock_cnt() == 1                      # full lock still held
+    assert 701 in t.drain_freed_swa_slots()
+    t.dec_lock_ref(n, b, True)                         # skip_swa
+    assert n.get_lock_cnt() == 0
+
+
+@cext
+def test_cpp_swa_locked_node_not_full_evictable():
+    """in_use() includes swa_lock_ref: SWA-locked node survives full eviction."""
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4], 0)
+    t.set_swa(n, 5)
+    n.inc_swa_lock_ref()
+    ev = _evict_full(t, 2)
+    assert len(ev) == 0            # locked -> not evicted
+    assert not t.is_empty()
+    n.dec_swa_lock_ref()
+
+
+@cext
+def test_cpp_reset_clears_tree():
+    t = _tree()
+    n = _insert(t, [1, 2, 3, 4], 0)
+    t.set_swa(n, 900)
+    t.reset()
+    assert t.is_empty()
+
+
+# =========================================================================== #
+# 3. WORKLOAD STRESS — production CacheEngineAccel + real SWA host-pool IO     #
+#    (folded in from the former benchmarks/benchmark_swa_nodemount.py)         #
+# =========================================================================== #
+#
+# A stochastic multi-turn dialogue workload that drives the SHIPPED path
+# (CacheEngineAccel: CRadixTreeIndex Full tree + real SWAHostPool byte IO)
+# under pool pressure, then asserts the aggregate two-pool lock-step invariants
+# that the per-op unit tests above cannot: no double-free, no use-after-free of
+# a live window, byte-identical SWA read-back, and no slot leak after reset.
+
+_WTPB = 16  # workload tokens_per_block (== swa_page_size: one page == one slot)
+
+
+@dataclass
+class _Turn:
+    token_ids: np.ndarray
+
+
+def _make_workload(num_users, num_turns, sys_len, turn_in, turn_out, seed):
+    rng = random.Random(seed)
+    npr = np.random.RandomState(seed)
+    vocab = 30000
+    system = npr.randint(0, vocab, size=sys_len, dtype=np.int64)
+    hist: Dict[int, np.ndarray] = {u: system.copy() for u in range(num_users)}
+    per_user_turns = {u: max(1, rng.randint(num_turns // 2, num_turns))
+                      for u in range(num_users)}
+    counters = {u: 0 for u in range(num_users)}
+    pending = list(range(num_users))
+    work: List[_Turn] = []
+    while pending:
+        u = rng.choice(pending)
+        inp = npr.randint(0, vocab, size=turn_in, dtype=np.int64)
+        prefix = np.concatenate([hist[u], inp])
+        work.append(_Turn(token_ids=prefix))
+        out = npr.randint(0, vocab, size=turn_out, dtype=np.int64)
+        hist[u] = np.concatenate([prefix, out])
+        counters[u] += 1
+        if counters[u] >= per_user_turns[u]:
+            pending.remove(u)
+    return work
+
+
+def _page_fingerprint(block_hash, slot_size):
+    """Deterministic bytes keyed by the window's block hash — stands in for the
+    real SWA-page KV; byte-identical read-back proves the slot round-tripped."""
+    rs = np.random.RandomState(block_hash & 0xFFFFFFFF)
+    return rs.randint(0, 256, size=slot_size, dtype=np.uint8)
+
+
+class _SWAWorkloadDriver:
+    """Minimal port of the former benchmark Driver: match(get) -> store(put)
+    with an SWA page mounted on each stored tail node, real byte IO on hit, and
+    ground-truth pool invariant sweeps."""
+
+    def __init__(self, num_cpu_blocks, swa_slots, slot_bytes):
+        from flexkv.cache.cache_engine import CacheEngineAccel
+        from flexkv.common.config import SWAPoolConfig
+        from flexkv.common.transfer import DeviceType
+        self.tpb = _WTPB
+        self.engine = CacheEngineAccel(
+            DeviceType.CPU, num_cpu_blocks, _WTPB,
+            evict_ratio=0.1, hit_reward_seconds=0,
+            evict_start_threshold=1.0, eviction_policy="lru",
+        )
+        self.engine.init_swa(SWAPoolConfig(
+            enabled=True, num_slots=swa_slots, window_size=_WTPB,
+            num_swa_layers=1, bytes_per_token_per_layer=max(1, slot_bytes // _WTPB),
+        ))
+        self.slot_size = self.engine.swa_pool.slot_size_bytes
+        self._slot_expect: Dict[int, int] = {}
+        self.violations: List[str] = []
+        self.swa_hits = 0
+        self.bytes_verified = 0
+
+    def _seq(self, token_ids):
+        aligned = (len(token_ids) // self.tpb) * self.tpb
+        sm = SequenceMeta(token_ids=token_ids[:aligned].astype(np.int64),
+                          tokens_per_block=self.tpb)
+        sm.gen_hashes()
+        return sm
+
+    def _reconcile(self):
+        free = set(int(s) for s in self.engine.swa_pool._free_slots)
+        for s in list(self._slot_expect):
+            if s in free:
+                self._slot_expect.pop(s, None)
+
+    def step(self, turn):
+        sm = self._seq(turn.token_ids)
+        nblocks = sm.num_blocks
+        if nblocks == 0:
+            return
+        mr = self.engine.match(sm)
+        full_hit = int(mr.num_ready_matched_blocks)
+        if full_hit > 0:
+            swa_hit, slot, _key = self.engine.match_swa(
+                sm, upper_bound_blocks=full_hit, lock_for_load=False)
+            if swa_hit > 0 and slot >= 0:
+                self.swa_hits += 1
+                got = np.asarray(self.engine.swa_pool.read_copy(slot)).ravel()
+                exp_hash = self._slot_expect.get(int(slot))
+                if exp_hash is not None:
+                    exp = _page_fingerprint(exp_hash, self.slot_size)
+                    if got.shape == exp.shape and np.array_equal(got, exp):
+                        self.bytes_verified += 1
+                    else:
+                        self.violations.append(f"SWA byte mismatch on slot {slot}")
+
+        num_new = nblocks - full_hit
+        if num_new > 0:
+            protected = mr.last_node if full_hit > 0 else None
+            phys_new = np.asarray(
+                self.engine.take(num_new, protected_node=protected, strict=False)
+            ).astype(np.int64)
+            self._reconcile()
+            if phys_new.size < num_new:
+                self.engine.recycle(phys_new)
+                return
+            node = self.engine.insert(sm, phys_new, num_insert_blocks=num_new,
+                                      is_ready=True, match_result=mr)
+        else:
+            node = mr.last_node
+        if node is None:
+            return
+        slot = self.engine.swa_alloc_slot()
+        self._reconcile()
+        if slot == -1:
+            return
+        tail_hash = int(sm.block_hashes[nblocks - 1])
+        self.engine.swa_pool.write(slot, _page_fingerprint(tail_hash, self.slot_size))
+        self._slot_expect[int(slot)] = tail_hash
+        self.engine.set_swa(node, slot)
+        self._reconcile()
+
+    def check_invariants(self):
+        pool = self.engine.swa_pool
+        used, free, total = pool.num_used, pool.num_free, pool.num_slots
+        if used + free != total:
+            self.violations.append(f"pool accounting {used}+{free}!={total}")
+        free_list = list(pool._free_slots)
+        if len(free_list) != len(set(free_list)):
+            self.violations.append("DOUBLE-FREE on the SWA free-list")
+        free_set = set(free_list)
+        stale = [s for s in self._slot_expect if s in free_set]
+        if stale:
+            self.violations.append(f"USE-AFTER-FREE: live slot(s) {stale[:8]} free")
+
+    def run(self, work, check_every=50):
+        for i, turn in enumerate(work):
+            self.step(turn)
+            if check_every and (i + 1) % check_every == 0:
+                self.check_invariants()
+        self.check_invariants()
+        self.engine.reset()
+        pool = self.engine.swa_pool
+        if pool.num_free != pool.num_slots:
+            self.violations.append(
+                f"LEAK after reset: free {pool.num_free} != total {pool.num_slots}")
+        return self.violations
+
+
+@cext
+@pytest.mark.parametrize("cpu_blocks,swa_slots", [
+    (4096, 256),   # baseline: little pressure
+    (512, 512),    # full-KV eviction pressure (I1 connect-free hot path)
+    (8192, 32),    # SWA-LRU pressure (I2 interior-first eviction hot path)
+])
+def test_workload_pressure_no_leak_no_corruption(cpu_blocks, swa_slots):
+    """Stochastic multi-turn workload on the production path: under Full and SWA
+    pool pressure, the two pools stay in lock-step — no double-free, no
+    use-after-free of a live window, byte-identical SWA read-back, no leak after
+    reset. (Folded from benchmark_swa_nodemount.py; small params for CI.)"""
+    work = _make_workload(num_users=24, num_turns=6, sys_len=8 * _WTPB,
+                          turn_in=6 * _WTPB, turn_out=2 * _WTPB, seed=1234)
+    drv = _SWAWorkloadDriver(cpu_blocks, swa_slots, slot_bytes=4096)
+    violations = drv.run(work, check_every=25)
+    assert not violations, "invariant violations:\n" + "\n".join(violations[:20])
+    # Sanity: the workload actually exercised SWA hits + real byte verify (so a
+    # silent no-op regression that stops mounting/reading SWA is caught).
+    assert drv.swa_hits > 0, "workload produced no SWA hits — coverage regressed"
+    assert drv.bytes_verified > 0, "no SWA byte round-trip verified"
+
+
+if __name__ == "__main__":
+    import sys
+    sys.exit(pytest.main([__file__, "-v"]))
